+++ console2/src/state/data/processes/types.ts
@@ -31,7 +31,8 @@ import { ProcessEntry } from '../../../api/process';
 import { RequestState } from '../common';
 import { State as LogState } from './logs/types';
 import { State as PollState } from './poll/types';
-import { State as historyState } from './history/types';
+import { State as HistoryState } from './history/types';
+import { State as AttachmentState } from './attachments/types';
 import { State as ChildrenState } from './children/types';
 import { State as EventsState } from './events/types';
 import { State as AnsibleState } from './ansible/types';
@@ -110,7 +111,8 @@ export interface State {
     ansible: AnsibleState;
     log: LogState;
     poll: PollState;
-    history: historyState;
+    history: HistoryState;
+    attachments: AttachmentState;
     children: ChildrenState;
     events: EventsState;
 }